Revert "Telas Home Cliente e Profissional "
